vlm_caption/requirements_vlm.txt
transformers>=4.30.0
accelerate
bitsandbytes
pillow
torch>=2.0.0
